# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A e-Commerce API in Go using Gin framework",
        "title": "Project Device Mart API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Anaz Ibinu Rasheed",
            "url": "https://www.linkedin.com/in/anaz-ibinu-rasheed-a2b461253/",
            "email": "anazibinurasheed@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/referral/claim": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Apply a  referral code  to get wallet money,",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "referral"
                ],
                "summary": "Apply referral code for referral bonus.",
                "parameters": [
                    {
                        "description": "Referral code to apply",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success, bonus amount updated in wallet.",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid Input.",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Failed.",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Failed.",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/referral/get-code": {
            "get": {
                "description": "Get the referral code of the currently logged-in user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "referral"
                ],
                "summary": "Get the referral code for the current user",
                "responses": {
                    "200": {
                        "description": "Success.",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Failed.",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "statuscode": {
                    "type": "integer"
                }
            }
        }
    }
}